// src/modules/movies/interfaces/Movie.ts
export interface Movie {
    title: string;
    plot: string;
    year: string;
    director: string;
    actors: string[];
    posterUrl: string;
    trailerUrl: string;
    boxOffice: string;
    released: string;
    writer: string;
    runtime: string;
    ratingImdb: string;
    imdbId: string;
    rated: string;
    genres: string[];
    _id?: string;
    magnet: string;
    sourceUrl: string;
    filename: string;
}